Adapt the `phonegap serve` command to work with CORS
In order to support iOS requirements for WKWebView, we use the cordova webview
in the app. This means that the HTML is now loaded from `http://localhost` on
android and `ionic://app` on iOS. This means that the calls to the phonegap
server actually use CORS.

The changes in this PR reflect the changes required to support serving content
over CORS. Notably:
- we add headers to support CORS origins, existing headers and allowing clients
  to read the header we use to exchange cookies
- implement a hacked-together version of cookies using local storage since
  cookies are not supported for CORS requests
  https://stackoverflow.com/questions/46288437/set-cookies-for-cross-origin-requests
  This involves changes to both the added client script (res_autoreload.js) and
  the server script (lib_autoreload.js)

Note that we implement these changes by copying the modified scripts into
`node_modules`. This is hacky in itself - we should really submit the fixes to
the `connect-phonegap` module instead.

But `connect-phonegap` has EOLed
https://www.npmjs.com/package/connect-phonegap
so that option is closed to us.

We should really move to an option other than cordova/phonegap, or if we want
to keep using it, fork the project ourselves.